Add logging (#49)
Fixes #46 

- set up basic logging

- default level is set to "INFO", but all logging is done at the "DEBUG"
level. So to see any of the messages we will need to set
`logger.setLevel("DEBUG")` in `src/napari_epitools/__init__.py`. I've
done this to hide these messages from users

- Any info, warning, or error messages that should be displayed to users
don't use the logger - instead they're shown using
`napari.utils.notifications`, which will show the message in a banner in
the gui rather than printing to the console

- some of the log messages have an even lower level than "DEBUG",
meaning they won't be printed when the log level is set to "DEBUG". This
is because it's likely these messages will be printed a lot, e.g. when
scrolling through frames, and they might make it difficult to see other
log messages in the console. To print all messages to the console:
`logger.setLevel(1)`