Colour 'Labels' layers by cell statistics (#53)
Fixes #51 



https://user-images.githubusercontent.com/29753790/230074652-1bf638b0-7e98-47fa-8288-bbf9f8479aef.mp4


- set minimum version of `scikit-image>=0.20`, as part of the API that
we're using change in this version (`skimage.future.RAG` was renamed to
`skimage.graph.RAG`). Also remove unused dependencies

- add widgets for creating colourmaps based on cell statistics

- to create a colourmap: select a `Labels` layer to create the
colourmaps for, set the range to use for creating the colourmaps, then
click 'Create colourmaps'

- this will create a series of colourmaps for the selected statistic -
one per frame - and store them as a list in
`Labels.metadata["colourmaps"]`

- when scrolling through frames, the colourmap is updated for the
current frame

- current options for colouring by statistic:
- id - in which case the `Labels.color_mode` is set to `auto`, which
will colour the cells by id
    - area
    - perimeter
    - orientation
    - number of neighbours

- there is an `Auto-limits` option for setting the range for the
colourmaps. If selected, the minimum and maximum values of the selected
statistic (across all frames) will be used for setting the range. When
creating the colourmap, the 'lower limit' and 'upper limit' widgets will
be updated with the automatically calculated values so the user know
what values have been used

- there are some issues / discussion with colouring a `Labels` layer by
features in `napari` that we might to want keep track of:
    - https://github.com/napari/napari/issues/4384
    - https://github.com/napari/napari/issues/5129
-
https://napari.zulipchat.com/#narrow/stream/212875-general/topic/labels.20.28direct.29
-
https://forum.image.sc/t/exloring-image-segment-features-in-napari/75222/3

- there are some other plugins that create colourmaps based on cell
statistics, but I couldn't find any that allow setting the colourmap
range. Plus, they tend to do the `regionprops` calculation on-the-fly
when the frame is changed rather than storing results for all frames in
the metadata, which means there is usually a lag when scrolling through
frames while the data is processed and colourmaps created

---------

Co-authored-by: Patrick Roddy <patrickjamesroddy@gmail.com>