Read pixel spacing from tif files and use when calculating regionprops (#58)
Fixes #52 

- add `PartSeg` as a dependency - it reads several file formats along
with some useful metadata

- PartSeg comes with napari readers for various file formats. However,
it [scales the
image](https://napari.org/stable/api/napari.layers.Image.html#:~:text=\)%20%E2%80%93%20Layer%20metadata.-,scale,-\(tuple%20of)
in the viewer based on the pixel spacing. This would be fine, but after
projecting the 3d image to 2d it is very difficult to fine to correct
plane that the 2d image is in (because the pixel spacing in `z` much
larger than the spacing in `xy`). So I've updated the EpiTools reader to
be a wrapper around one of the PartSeg readers that removes the scaling
from

- use the pixel spacing when calculating cell statistics with
regionprops. Area and perimeter are then converted to be in
$\micro\text{m}^2$ and $\micro\text{m}^2$, respectively