Add everything from `regionprops` (#93)
Fixes #90.

Researchers mentioned that it would be useful to have everything from
`regionprops` rather than having only certain things.

There was some discussion with units in #52. However, the researchers
thought it would be easier to just use the `regionprops` defaults.

See #96, #97, #98 & #100 for more detail on some of the changes.

Happy to discuss. This was more complicated than I thought it would be.